Add check state method
splitting out set_state and check_state means check state can be called
after an animation loop has finished. This allows one anim to run once and
then be followed by a different animation or action. An example: player
is hurt and the hurt animation plays. Following this, if the player's
health is reduced to zero (or less) the death animation plays and a
death() method on the player can be run.